新增、修改已去重

+++ src/main/java/com/zy_admin/sys/service/impl/SysRoleServiceImpl.java
@@ -15,7 +15,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
-import javax.jws.soap.SOAPBinding;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -87,6 +86,7 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleDao, SysRole> impleme
 
     /**
      * 新增角色
+     *
      * @param roleAndRoleMenu
      * @return
      */
@@ -94,24 +94,17 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleDao, SysRole> impleme
     @Override
     @Transactional(rollbackFor = Exception.class)
     public Result insert(RoleAndRoleMenu roleAndRoleMenu) {
+        //判断角色名和权限标识是否重复
         String roleName = roleAndRoleMenu.getRoleName();
         String roleKey = roleAndRoleMenu.getRoleKey();
         SysRole checkRoleName = this.baseMapper.selectRoleName(roleName);
         SysRole checkRoleKey = this.baseMapper.selectRoleKey(roleKey);
-        System.out.println("N-----");
-        System.out.println(roleName);
-        System.out.println(checkRoleName);
-
-        System.out.println("K----");
-        System.out.println(roleKey);
-        System.out.println(checkRoleKey);
-
-        if (checkRoleName==null)
-        {
-            if (checkRoleKey==null)
-            {
+        if (checkRoleName == null) {
+            if (checkRoleKey == null) {
                 try {
+                    //插入角色
                     this.baseMapper.insert(roleAndRoleMenu);
+                    //判断是否有权限，有则插入
                     if (roleAndRoleMenu.getMenuIds().length != 0) {
                         sysRoleMenuDao.insertBatch(roleAndRoleMenu.getRoleId(), roleAndRoleMenu.getMenuIds());
                     }
@@ -120,32 +113,47 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleDao, SysRole> impleme
                     e.printStackTrace();
                     return new Result(null, ResultTool.fail(ResultCode.COMMON_FAIL));
                 }
-            }else {
-                return new Result(null,ResultTool.fail(ResultCode.REPEAT_ROLE_KEY));
+            } else {
+                return new Result(null, ResultTool.fail(ResultCode.REPEAT_ROLE_KEY));
             }
-
-        }else {
-            return new Result(null,ResultTool.fail(ResultCode.REPEAT_ROLE_NAME));
+        } else {
+            return new Result(null, ResultTool.fail(ResultCode.REPEAT_ROLE_NAME));
         }
-
-
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public Result update(RoleAndRoleMenu roleAndRoleMenu) {
         Result result = new Result();
-        int i = this.baseMapper.updateRoleById(roleAndRoleMenu);
-        if (i == 1) {
-            sysRoleMenuDao.deleteById(roleAndRoleMenu.getRoleId());
-            if (roleAndRoleMenu.getMenuIds().length != 0) {
-                sysRoleMenuDao.insertBatch(roleAndRoleMenu.getRoleId(), roleAndRoleMenu.getMenuIds());
+        String roleName = roleAndRoleMenu.getRoleName();
+        String roleKey = roleAndRoleMenu.getRoleKey();
+        SysRole checkRoleName = this.baseMapper.selectRoleName(roleName);
+        SysRole checkRoleKey = this.baseMapper.selectRoleKey(roleKey);
+        if (checkRoleName == null) {
+            if (checkRoleKey == null) {
+                try {
+                    //更新角色
+                    int i = this.baseMapper.updateRoleById(roleAndRoleMenu);
+                    if (i == 1) {
+                        //判断是否有权限，有则更改
+                        if (roleAndRoleMenu.getMenuIds().length != 0) {
+                            sysRoleMenuDao.deleteById(roleAndRoleMenu.getRoleId());
+                            sysRoleMenuDao.insertBatch(roleAndRoleMenu.getRoleId(), roleAndRoleMenu.getMenuIds());
+                        }
+                        result.setMeta(ResultTool.fail(ResultCode.SUCCESS));
+                        return result;
+                    }
+                    return new Result(null, ResultTool.fail(ResultCode.SUCCESS));
+                } catch (Exception e) {
+                    e.printStackTrace();
+                    return new Result(null, ResultTool.fail(ResultCode.COMMON_FAIL));
+                }
+            } else {
+                return new Result(null, ResultTool.fail(ResultCode.REPEAT_ROLE_KEY));
             }
-            result.setMeta(ResultTool.fail(ResultCode.SUCCESS));
-            return result;
+        } else {
+            return new Result(null, ResultTool.fail(ResultCode.REPEAT_ROLE_NAME));
         }
-        result.setMeta(ResultTool.fail(ResultCode.COMMON_FAIL));
-        return result;
     }
 
     @Override
